+++ docs/requirements.md
@@ -19,8 +19,9 @@
 
 ### UI
 
-[ ] 영화 카드는 하나의 컴포넌트로 구성한다.\
-[ ] 한 페이지당 20개씩 영화 목록을 보여준다.\
-[ ] 각 페이지의 최하단에는 더보기 버튼이 있다.\
+[x] 영화 카드는 하나의 컴포넌트로 구성한다.\
+[x] 한 페이지당 20개씩 영화 목록을 보여준다.\
+[x] 각 페이지의 최하단에는 더보기 버튼이 있다.\
+[x] 더보기 버튼을 누르면 영화 목록을 추가하여 보여준다.\
 [ ] 마지막 페이지에는 더보기 버튼이 없다.\
 [ ] 영화 목록 아이템에 대한 Skeleton UI를 구현한다.
