+++ index.html
@@ -19,7 +19,7 @@
         <section class="item-view">
           <h2>지금 인기 있는 영화</h2>
           <ul class="item-list"></ul>
-          <button class="btn primary full-width">더 보기</button>
+          <button class="show-more btn primary full-width">더 보기</button>
         </section>
       </main>
     </div>
