+++ src/index.js
@@ -1,15 +1,21 @@
+import "../templates/logo.png";
+import "../templates/star_filled.png";
 import App from "./js/domain/App.js";
 import { $ } from "./utils/dom.js";
-import { generateMovieCard } from "./js/components/MovieCard.js";
-import "../templates/logo.png";
+import MovieCardList from "./js/view/MovieCardList.js";
 
 const app = new App();
 
 addEventListener("DOMContentLoaded", async () => {
   await app.init();
 
-  const movieList = $(".item-list");
+  MovieCardList.render(app.movies);
+});
 
-  const movieCards = app.movies.map((movie) => generateMovieCard(movie));
-  movieList.append(...movieCards);
+const showMoreButton = $(".show-more");
+
+showMoreButton.addEventListener("click", async () => {
+  await app.fetchNextPage();
+
+  MovieCardList.render(app.movies);
 });
