--- src/js/components/MovieCard.js
@@ -1,20 +1,0 @@
-function generateMovieCard(movie) {
-  const card = document.createElement("li");
-  card.innerHTML = /* html */ `
-  <a href="#">
-  <div class="item-card">
-    <img
-      class="item-thumbnail"
-      src=${movie.thumbnail}
-      loading="lazy"
-      alt=${movie.title}
-    />
-    <p class="item-title">${movie.title}</p>
-    <p class="item-score"><img src="./star_filled.png" alt="별점" /> ${movie.rating}</p>
-  </div>
-</a>
-  `;
-  return card;
-}
-
-export { generateMovieCard };
