+++ src/js/view/MovieCardList.js
@@ -1,0 +1,17 @@
+import MovieCard from "./MovieCard.js";
+
+const MovieCardList = {
+  elements: {
+    movieCardList: document.querySelector(".item-list"),
+  },
+  render: (movies) => {
+    const movieCards = movies.map((movie) =>
+      MovieCard.generateMovieCard(movie)
+    );
+
+    MovieCardList.elements.movieCardList.innerHTML = "";
+    MovieCardList.elements.movieCardList.append(...movieCards);
+  },
+};
+
+export default MovieCardList;
